requirements.txt
discord
discord-py-interactions
python-dotenv
beautifulsoup4
requests